requirements.txt
azure-common==1.0.0
azure-nspkg==1.0.0
azure-storage==0.20.1
Cython==0.23
decorator==4.0.2
matplotlib==1.4.3
networkx==1.10
nose==1.3.7
numexpr==2.4.3
numpy==1.9.2
pandas==0.17.0
py4j==0.9
pyparsing==2.0.3
python-dateutil==2.4.2
python-igraph==0.7.1.post6
pytz==2015.4
requests==2.7.0
scikit-learn==0.16.1
scipy==0.16.0
six==1.9.0
tables==3.2.2
wheel==0.24.0
